+++ src/Agents.jl
@@ -7,7 +7,7 @@ using Graphs
 using DataFrames
 using Random
 import ProgressMeter
-import Base.length
+import Base.length # TODO: This should not be imported!!!
 import LinearAlgebra
 
 # Core structures of Agents.jl
@@ -18,7 +18,9 @@ include("core/space_interaction_API.jl")
 # Existing spaces
 include("spaces/nothing.jl")
 include("spaces/graph.jl")
-include("spaces/grid.jl")
+include("spaces/grid_general.jl")
+include("spaces/grid_multi.jl")
+include("spaces/grid_single.jl")
 include("spaces/discrete.jl")
 include("spaces/continuous.jl")
 include("spaces/openstreetmap.jl")
@@ -42,33 +44,36 @@ include("deprecations.jl")
 
 # Update messages:
 using Scratch
-display_update = true
-version_number = "5"
-update_name = "update_v$(version_number)"
 
 function __init__()
-    if display_update
-        # Get scratch space for this package
-        versions_dir = @get_scratch!("versions")
-        if !isfile(joinpath(versions_dir, update_name))
-            printstyled(
-                stdout,
-                """
-                \nUpdate message: Agents v$(version_number)
-                Welcome to this new major version of Agents.jl!
-                Noteworthy changes:
+display_update = true
+version_number = "5.4"
+update_name = "update_v$(version_number)"
+update_message = """
+Update message: Agents v$(version_number)
+Welcome to this massive update of Agents.jl!
+Noteworthy changes:
 
-                * Schedulers have been reworked to be more performant and allocate
-                  less. This means that most scheduler names have been deprecated from
-                  functions to types, such as `by_type` -> `ByType`. See changelog for
-                  full list!
-                * See the CHANGELOG.md or online docs for more!
-                """;
-                color=:light_magenta
-            )
-            touch(joinpath(versions_dir, update_name))
-        end
+- About 5x performance increase in distributed computing!
+- Internals of `GridSpace` have been completely re-written! This led to a significant
+  performance increase of about 30%!
+- New space `GridSpaceSingle` that is the same as `GridSpace` but only allows for one
+  agent per position only. It utilizes this knowledge for massive performance benefits
+  over `GridSpace`, **being about 3x faster than `GridSpace`** all across the board!
+- Performance increase for nearby_stuff searches `ContinuousSpace` (2-5x faster)
+- New `:manhattan` metric for `GridSpace` models!
+- Several new deprecations and/or possible breaking changes!
+
+See the CHANGELOG.md or online docs for more!
+"""
+if display_update
+    # Get scratch space for this package
+    versions_dir = @get_scratch!("versions")
+    if !isfile(joinpath(versions_dir, update_name))
+        printstyled(stdout, "\n"*update_message; color=:light_magenta)
+        touch(joinpath(versions_dir, update_name))
     end
 end
+end # _init__ function.
 
 end # module